Update closed_loop_eval.py
New episodes are generated EVERY TIME the python file is ran. No 2 runs of the python file will produce the same result

BUT modified the file to use the same episodes for all models being tested for fairness. For example, if i run the python file and test 5 different models, the file now generates a unique set of episodes and tests all 5 models with this set of episodes. This is to ensure fair comparison, and that the models are not doing better/worse because they got easier/harder episodes